float

+++ hello_world/6-size.c
@@ -10,6 +10,6 @@ printf("Size of char: %lu bytes\n", sizeof(unsigned char));
 printf("Size of an int: %lu bytes\n", sizeof(unsigned int));
 printf("Size of a long int: %lu bytes\n", sizeof(unsigned long int));
 printf("Size of a long long int: %lu bytes\n", sizeof(unsigned long long int));
-printf("Size of a float: %lu bytes\n", sizeof(unsigned float));
+printf("Size of a float: %lu bytes\n", sizeof(float));
 return (0);
 }
